youtube/servers/py-mcp-youtube-toolbox/pyproject.toml
[project]
name = "py-mcp-youtube-toolbox"
version = "0.1.0"
description = "YouTube Toolbox"
authors = [
    { name = "Anthony Kim", email = "jikime@gmail.com" },
]
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.169.0",
    "mcp[cli]>=1.7.1",
    "python-dotenv>=1.1.0",
    "youtube-transcript-api>=1.0.3",
]
